create separate MQL Visitor class
