+++ snuba_sdk/metrics_query.py
@@ -6,7 +6,8 @@ from typing import Any, Mapping
 
 from snuba_sdk.expressions import Limit, Offset
 from snuba_sdk.formula import Formula
-from snuba_sdk.metrics_query_visitors import MQLPrinter, SnQLPrinter, Validator
+from snuba_sdk.metrics_query_visitors import SnQLPrinter, Validator
+from snuba_sdk.mql_visitor import MQLPrinter
 from snuba_sdk.query import BaseQuery
 from snuba_sdk.query_visitors import InvalidQueryError
 from snuba_sdk.timeseries import MetricsScope, Rollup, Timeseries
